# server/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EaaS Nexus Backend API",
    "version": "1.0.0",
    "description": "Backend API contract for EaaS Nexus mobile application."
  },
  "servers": [
    { "url": "https://server-plum-six.vercel.app", "description": "Production" },
    { "url": "http://localhost:3001", "description": "Development" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": { "type": "object", "additionalProperties": true }
        },
        "required": ["success"]
      }
    }
  },
  "security": [{ "bearerAuth": [] }],
...
  "paths": {
...
    "/notifications": {
      "get": {
        "tags": ["Notifications"],
        "summary": "Get user notifications",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": { "type": "integer", "default": 0 }
          },
          {
            "in": "query",
            "name": "unreadOnly",
            "schema": { "type": "boolean", "default": false }
          }
        ],
        "responses": {
          "200": {
            "description": "Notifications",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/notifications/{notificationId}/read": {
      "patch": {
        "tags": ["Notifications"],
        "summary": "Mark notification as read",
        "parameters": [
          {
            "in": "path",
            "name": "notificationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Marked read",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/notifications/admin/broadcast": {
      "post": {
        "tags": ["Notifications"],
        "summary": "Admin broadcast notification",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "title": "System update",
                "message": "Planned maintenance window tonight.",
                "type": "announcement",
                "route": "/dashboard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Broadcast sent",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/support/ticket": {
      "post": {
        "tags": ["Support"],
        "summary": "Create support ticket",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": {
                "propertyId": "prop_001",
                "category": "technical",
                "priority": "high",
                "title": "Solar panel issue",
                "description": "My solar panel stopped generating power."
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ticket created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
    "/support/admin/tickets": {
      "get": {
        "tags": ["Support"],
        "summary": "Admin list support tickets",
        "parameters": [
          { "in": "query", "name": "status", "schema": { "type": "string" } },
          {
            "in": "query",
            "name": "limit",
            "schema": { "type": "integer", "default": 50 }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": { "type": "integer", "default": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin tickets",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
    "/support/admin/tickets/{ticketId}": {
      "patch": {
        "tags": ["Support"],
        "summary": "Admin update support ticket",
        "parameters": [
          {
            "in": "path",
            "name": "ticketId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "example": { "status": "in_progress", "priority": "high" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Ticket updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SuccessResponse" }
              }
            }
          }
        }
      }
    },
...
  }
}